pyproject.toml
[project]
name = "ghost-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "fastapi[standard]>=0.129.0",
    "psycopg2>=2.9.11",
    "psycopg[binary,pool]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.46",
    "websockets>=16.0",
]
